online_inference/requirements.txt
numpy~=1.23.1
pandas==1.5.0
scikit-learn==1.0.2
uvicorn==0.19.0
pyyaml==6.0
pytest==7.1.3
click~=8.0.4
marshmallow-dataclass==8.5.9
fastapi==0.85.1
requests==2.28.1
